+++ akd/tools/search/utils.py
@@ -11,8 +11,10 @@ from __future__ import annotations
 import asyncio
 import re
 from collections import defaultdict
+from typing import Callable
 from urllib.parse import urlparse
 
+import numpy as np
 from loguru import logger
 from pydantic import AnyUrl
 
@@ -329,6 +331,189 @@ def sort_results(
         if debug:
             logger.warning(f"Sorting by {sort_by} failed due to mixed types.")
         return results
+
+
+def reciprocal_rank_fusion(
+    *results: list["SearchResultItem"],
+    k: int = 60,
+    keys: list[str] | str | None = None,
+    value_normalizers: dict[str, Callable] | None = None,
+    normalize: bool = True,
+    debug: bool = False,
+) -> list["SearchResultItem"]:
+    """
+    Fuse multiple ranked lists using Reciprocal Rank Fusion (RRF) with multi-key matching and value normalization.
+
+    Formula: RRF_score(item) = Σ(1 / (k + rank)) across all lists containing item
+
+    Args:
+        results: Variable number of ranked result lists (one per query).
+        k: RRF constant (default 60, standard value from literature).
+        keys: List of attribute names for deduplication (cascaded OR logic).
+              Default: ["doi", "title", "url"] - matches if ANY key matches.
+              Can also pass single string for backward compatibility.
+        value_normalizers: Optional dict mapping key names to normalizer functions.
+              Each function takes a SearchResultItem and returns a normalized string or None.
+              Default normalizers: doi (extract from URL), title (lowercase, no punctuation), url (no protocol/www).
+              Pass empty dict {} to disable normalization.
+              Example: {"url": lambda item: my_custom_normalizer(item.url)}
+        normalize: If True, apply min-max normalization to scores [0.1, 1.0] (default True).
+                  Raw RRF scores are always preserved in item.extra["rrf_score"].
+        debug: Enable debug logging showing matched keys and values.
+
+    Returns:
+        Fused list sorted by RRF score (descending). Original item values are preserved.
+
+    Notes:
+        - Normalization is applied ONLY for matching/deduplication, not for output values
+        - Items with URLs "https://github.com/foo" and "https://www.github.com/foo/" will match
+        - Items with titles "Deep Learning: A Survey" and "deep learning a survey" will match
+        - The first encountered item's original values are kept in the output
+
+    Examples:
+        >>> # Default normalization (recommended)
+        >>> results_q1 = [item_a, item_b, item_c]
+        >>> results_q2 = [item_b, item_d, item_a]
+        >>> fused = reciprocal_rank_fusion(results_q1, results_q2)
+        # item_b and item_a appear in both, so they get boosted
+
+        >>> # URL normalization matches different URL formats
+        >>> item1 = SearchResultItem(url="https://github.com/foo/bar")
+        >>> item2 = SearchResultItem(url="https://www.github.com/foo/bar/")
+        >>> fused = reciprocal_rank_fusion([item1], [item2], keys=["url"])
+        # Result: 1 item (matched via normalized URL), original URL preserved
+
+        >>> # Custom normalizer
+        >>> custom = {"url": lambda item: item.url.lower().strip()}
+        >>> fused = reciprocal_rank_fusion(results_q1, results_q2, value_normalizers=custom)
+
+        >>> # Disable normalization
+        >>> fused = reciprocal_rank_fusion(results_q1, results_q2, value_normalizers={})
+    """
+    # Handle default and backward compatibility
+    keys = keys or ["doi", "title", "url"]
+    if isinstance(keys, str):
+        keys = [keys]
+
+    # Return empty if no results
+    if not results:
+        return []
+
+    # Track: identifier -> cumulative RRF score
+    # Track: identifier -> SearchResultItem
+    # Track: identifier -> set of all (key, value) identifiers for this item
+    identifier_to_score = defaultdict(float)
+    identifier_to_item = {}
+    identifier_groups = defaultdict(set)
+
+    # Import normalizer functions (avoid circular import at module level)
+
+    # Use provided normalizers or defaults
+    value_normalizers = value_normalizers or {
+        "doi": lambda item: get_doi(item),
+        "title": lambda item: normalize_title(item.title) if item.title else None,
+        "url": lambda item: normalize_url(item.url) if item.url else None,
+    }
+
+    # Calculate RRF scores with multi-key matching
+    for rank_list in results:
+        for rank, item in enumerate(rank_list, 1):
+            # Collect all non-None identifiers for this item (using normalized values)
+            item_identifiers = set()
+            for key in keys:
+                # Apply normalizer if available, otherwise use original value
+                if key in value_normalizers:
+                    value = value_normalizers[key](item)
+                else:
+                    value = getattr(item, key, None)
+
+                if value:
+                    item_identifiers.add((key, str(value)))
+
+            if not item_identifiers:
+                continue  # Skip items with no valid identifiers
+
+            # Pick canonical identifier (prefer first key in priority order)
+            canonical = None
+            for key in keys:
+                for item_key, item_val in item_identifiers:
+                    if item_key == key:
+                        canonical = (item_key, item_val)
+                        break
+                if canonical:
+                    break
+
+            if not canonical:
+                canonical = next(iter(item_identifiers))
+
+            # Check if any identifier matches existing groups
+            matched_canonical = None
+            for existing_canonical, group in identifier_groups.items():
+                if item_identifiers & group:  # Set intersection - shared identifier?
+                    matched_canonical = existing_canonical
+                    break
+
+            increment = 1.0 / (rank + k)
+            if matched_canonical:
+                # Merge with existing item - accumulate RRF score
+                identifier_to_score[matched_canonical] += increment
+
+                if debug:
+                    # Show which keys matched (BEFORE updating the group!)
+                    shared = item_identifiers & identifier_groups[matched_canonical]
+                    matched_keys = [k for k, _ in shared]
+                    canonical_key, canonical_val = matched_canonical
+                    logger.debug(
+                        f"[RRF] MERGED: rank={rank} matched via {matched_keys} | "
+                        f"primary_key={canonical_key} value='{canonical_val[:40]}...' | +score={increment:.6f}",
+                    )
+
+                # Merge all identifiers into the group (AFTER logging!)
+                identifier_groups[matched_canonical].update(item_identifiers)
+            else:
+                # New item - create new group
+                identifier_to_score[canonical] += increment
+                identifier_to_item[canonical] = item.model_copy()
+                identifier_groups[canonical] = item_identifiers
+                if debug:
+                    available_keys = [k for k, _ in item_identifiers]
+                    canonical_key, canonical_val = canonical
+                    logger.debug(
+                        f"[RRF] NEW: rank={rank} | "
+                        f"primary_key={canonical_key} value='{canonical_val[:40]}...' | "
+                        f"available_keys={available_keys}",
+                    )
+
+    # Build results with RRF scores
+    fused_results = []
+    for identifier, rrf_score in sorted(identifier_to_score.items(), key=lambda x: x[1], reverse=True):
+        item = identifier_to_item[identifier]
+        item.score = rrf_score
+        item.extra = item.extra or {}
+        item.extra["rrf_score"] = rrf_score
+        # Track which keys were used for matching (useful for debugging)
+        item.extra["rrf_matched_keys"] = list(set(k for k, _ in identifier_groups[identifier]))
+        fused_results.append(item)
+
+    if not fused_results:
+        return []
+
+    # Normalize scores if requested
+    if normalize:
+        scores = np.array([item.score for item in fused_results])
+        score_range = scores.max() - scores.min()
+
+        if score_range > 0:
+            # Scale to [0.1, 1.0] range to avoid 0.0 scores
+            normalized_scores = 0.1 + 0.9 * (scores - scores.min()) / score_range
+        else:
+            # All scores are the same
+            normalized_scores = np.ones_like(scores)
+
+        for item, norm_score in zip(fused_results, normalized_scores):
+            item.score = float(norm_score)
+
+    return fused_results
 
 
 __all__ = [
